refactoring cookies arquitecture
