+++ Web Ui/src/modules/Auth/application/deleteSessionCookie.ts
@@ -1,0 +1,10 @@
+import Cookies from 'js-cookie';
+
+export const removeSessionCookie = () => {
+  try {
+    Cookies.remove('userSession');
+    console.log('Session cookie removed successfully.');
+  } catch (error) {
+    console.error('Error removing session cookie:', error);
+  }
+};
